mainly some moving of things around
